Separate out uvh5_to_ms, refactoring
